feat: ledger with entry log per ledger
